This is the completed test for rubber cheese
****IMPROVEMENTS****
Using the latest version of Chrome there are a number of layout issues,
with some elements completely out of position. The paddings and margins
of some elements don't correspond to the design. The wrong font was also
used, it should have been Open Sans instead of Raleway. A lot of the
images were saved in an inappropriate file format - I would suggest
saving them as .PNG files instead of .GIFs.
I can also see an attempt to make the website responsive, which is
great. However, it's incomplete and doesn't fully work. I would also
suggest using percentages in the future as it makes it respond more
fluidly and makes it easier to develop responsive layouts - you don't
usually require as many breakpoints.